+++ src/startup_request.hpp
@@ -40,7 +40,7 @@ public:
   const std::string compression() const { return compression_; }
 
 private:
-  int32_t encode(int version, BufferValueVec* bufs) const;
+  ssize_t encode(int version, BufferValueVec* bufs) const;
 
 private:
   typedef std::map<std::string, std::string> OptionsMap;
